requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1

python-dotenv==0.20.0
requests>=2.31,<3

pandas>=2.1,<3
numpy>=1.26,<3
scikit-learn>=1.3,<2
folium>=0.15,<1

harmony-py>=1.3.1,<2

xarray>=2024.1
xarray-datatree>=0.0.14
netCDF4

pydantic[email]>=2.6,<3
email-validator>=2.1,<3
